rename util.constants package to util.constant
