+++ src/main/java/self/izouir/bitkionline/commander/HelpCommander.java
@@ -11,8 +11,8 @@ import java.util.List;
 
 import static self.izouir.bitkionline.util.BotMessageSender.deleteMessage;
 import static self.izouir.bitkionline.util.BotMessageSender.sendMessage;
-import static self.izouir.bitkionline.util.constants.ReplyMarkupConstants.CLOSE_BUTTON_TEXT;
-import static self.izouir.bitkionline.util.constants.commander.HelpCommanderConstants.HELP_MESSAGE;
+import static self.izouir.bitkionline.util.constant.ReplyMarkupConstant.CLOSE_BUTTON_TEXT;
+import static self.izouir.bitkionline.util.constant.commander.HelpCommanderConstant.HELP_MESSAGE;
 
 @Component
 public class HelpCommander {
