+++ src/main/java/self/izouir/bitkionline/util/constant/ReplyMarkupConstant.java
@@ -1,9 +1,9 @@
-package self.izouir.bitkionline.util.constants;
+package self.izouir.bitkionline.util.constant;
 
 import lombok.experimental.UtilityClass;
 
 @UtilityClass
-public class ReplyMarkupConstants {
+public class ReplyMarkupConstant {
     public static final String REFRESH_BUTTON_TEXT = "Refresh";
     public static final String YES_BUTTON_TEXT = "Yes";
     public static final String NO_BUTTON_TEXT = "No";
