+++ src/main/java/self/izouir/bitkionline/util/constant/service/PrivateBattleServiceConstant.java
@@ -1,9 +1,9 @@
-package self.izouir.bitkionline.util.constants.service;
+package self.izouir.bitkionline.util.constant.service;
 
 import lombok.experimental.UtilityClass;
 
 @UtilityClass
-public class PrivateBattleServiceConstants {
+public class PrivateBattleServiceConstant {
     public static final Integer LINK_LENGTH = 12;
     public static final String LINK_ALPHABET = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789#_";
 }
